Move Function Activator Validator to fix class path issue
