+++ legend-engine-xts-functionActivator/legend-engine-xt-functionActivator-generation/src/main/java/org/finos/legend/engine/functionActivator/validation/FunctionActivatorError.java
@@ -12,7 +12,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package org.finos.legend.engine.functionActivator.service;
+package org.finos.legend.engine.functionActivator.validation;
 
 public class FunctionActivatorError
 {
