+++ legend-engine-xts-functionActivator/legend-engine-xt-functionActivator-http-api/pom.xml
@@ -95,6 +95,14 @@
             <groupId>org.finos.legend.engine</groupId>
             <artifactId>legend-engine-language-pure-modelManager</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.finos.legend.engine</groupId>
+            <artifactId>legend-engine-xt-functionActivator-generation</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.finos.legend.engine</groupId>
+            <artifactId>legend-engine-xt-functionActivator-protocol</artifactId>
+        </dependency>
         <!-- ENGINE -->
 
         <!-- ANNOTATIONS -->
@@ -157,11 +165,6 @@
             <artifactId>log4j</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>org.finos.legend.engine</groupId>
-            <artifactId>legend-engine-xt-functionActivator-protocol</artifactId>
-        </dependency>
-
         <!-- TEST -->
     </dependencies>
 </project>
